Added explicit convertion to size_t

+++ src/Core/IO/TypeInfo.hh
@@ -336,11 +336,11 @@ namespace RavlN {
         return 0;
       }
 
-      SArray1dC<DataT> data(size);
+      SArray1dC<DataT> data(static_cast<size_t>(size));
       Int64T dataRead = iport.GetArray(data);
       if (dataRead > 0)
       {
-        SArray1dC<DataT> subData(data, dataRead);
+        SArray1dC<DataT> subData(data, static_cast<size_t>(dataRead));
         strm << subData;
       }
